tiliqua_platform: move top_level_cli to cli.py
