+++ .github/workflows/main.yml
@@ -13,7 +13,7 @@ jobs:
       - run: yosys --version
       - run: |
           pdm install
-          pdm build_usb_audio
+          pdm run src/example_usb_audio/top.py
         working-directory: gateware
       - uses: actions/upload-artifact@v4
         with:
@@ -84,7 +84,7 @@ jobs:
       - run: yosys --version
       - run: |
           pdm install
-          pdm build_xbeam
+          pdm run src/xbeam/top.py
         working-directory: gateware
       - uses: actions/upload-artifact@v4
         with:
@@ -108,7 +108,7 @@ jobs:
       - run: yosys --version
       - run: |
           pdm install
-          pdm build_selftest
+          pdm run src/selftest/top.py
         working-directory: gateware
       - uses: actions/upload-artifact@v4
         with:
@@ -133,7 +133,7 @@ jobs:
       - run: yosys --version
       - run: |
           pdm install
-          pdm build_poly
+          pdm run src/poly/top.py
         working-directory: gateware
       - uses: actions/upload-artifact@v4
         with:
@@ -150,7 +150,7 @@ jobs:
       - run: yosys --version
       - run: |
           pdm install
-          pdm build_vectorscope
+          pdm run src/example_vectorscope/top.py
         working-directory: gateware
       - uses: actions/upload-artifact@v4
         with:
@@ -174,7 +174,7 @@ jobs:
       - run: yosys --version
       - run: |
           pdm install
-          pdm build_boot
+          pdm run src/boot/top.py
         working-directory: gateware
       - uses: actions/upload-artifact@v4
         with:
@@ -192,9 +192,9 @@ jobs:
       - run: yosys --version
       - run: |
           pdm install
-          pdm build_boot_stub 0x100000 1
+          pdm run src/boot_stub/top.py --bootaddr 0x100000
         working-directory: gateware
       - uses: actions/upload-artifact@v4
         with:
           name: bootstub1.bit
-          path: gateware/build/bootstub1.bit
+          path: gateware/build/top.bit
