+++ gateware/src/boot/top.py
@@ -4,7 +4,8 @@
 
 import os
 
-from tiliqua.tiliqua_soc                 import TiliquaSoc, top_level_cli
+from tiliqua.tiliqua_soc                 import TiliquaSoc
+from tiliqua.cli                         import top_level_cli
 
 if __name__ == "__main__":
     this_path = os.path.dirname(os.path.realpath(__file__))
