+++ gateware/src/poly/top.py
@@ -17,7 +17,8 @@ from amaranth_future                             import fixed
 
 from tiliqua                                     import eurorack_pmod, dsp, midi
 from tiliqua.eurorack_pmod                       import ASQ
-from tiliqua.tiliqua_soc                         import TiliquaSoc, top_level_cli
+from tiliqua.tiliqua_soc                         import TiliquaSoc
+from tiliqua.cli                                 import top_level_cli
 
 from xbeam.top                                   import VectorTracePeripheral
